# docs/swagger.yaml
definitions:
...
  handler.PaymentRequest:
    properties:
      amount:
        type: number
      purpose:
        description: Either "Top-Up" or "Rental Payment"
        type: string
    required:
    - amount
    - purpose
    type: object
...
  handler.WalletBalanceResponse:
    properties:
      balance:
        type: number
    type: object
info:
  contact: {}
paths:
...
  /payment/create:
    post:
      consumes:
      - application/json
      description: Allows customers to create payment requests (e.g., Top-Up)
      parameters:
      - description: Payment request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment request details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a payment request
      tags:
      - Transactions
  /payment/status/{orderID}:
    get:
      description: Manually check the status of a payment using the order ID
      parameters:
      - description: Order ID
        in: path
        name: orderID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment status details
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check payment status
      tags:
      - Transactions
...
  /wallet/balance:
    get:
      description: Fetch the wallet balance of the authenticated customer
      produces:
      - application/json
      responses:
        "200":
          description: Customer's wallet balance
          schema:
            $ref: '#/definitions/handler.WalletBalanceResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get wallet balance
      tags:
      - Transactions
swagger: "2.0"